refactor: 데이터 fetch 함수를 커스텀 함수로 변경
