+++ src/pages/shift/ShiftPage.tsx
@@ -1,26 +1,22 @@
 import ShiftList from '@/components/list/ShiftList';
 import Select from '@/components/ui/Select';
 import { SELECT_APPROVAL_TYPE, SELECT_WORK_TYPE } from '@/constants/constant';
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import * as S from './ShiftPage.styles';
-import { getShiftList } from '@/api/shiftApi';
-import { IShiftList, ShiftListItem } from '@/types/shift';
-import { removeDuplicates } from '@/utils/arrayUtils';
+import { useShiftList } from '@/hooks/useShiftList';
 
 const ShiftPage = () => {
   const [workType, setWorkType] = useState('');
   const [approvalType, setApprovalType] = useState('');
-  const [page, setPage] = useState(0);
-  const [shiftList, setShiftList] = useState<ShiftListItem[]>([]);
-  const [isLoading, setIsLoading] = useState(false);
-  const [hasLoadMore, setHasLoadMore] = useState(true);
+  const { isLoading, shiftList, loadMore, resetList } = useShiftList({
+    workType,
+    approvalType,
+  });
 
   const handleWorkChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
     if (isLoading) return;
     setWorkType(event.target.value);
-    setShiftList([]);
-    setPage(0);
-    setHasLoadMore(true);
+    resetList();
   };
 
   const handleApprovalChange = (
@@ -28,43 +24,7 @@ const ShiftPage = () => {
   ) => {
     if (isLoading) return;
     setApprovalType(event.target.value);
-    setShiftList([]);
-    setPage(0);
-    setHasLoadMore(true);
-  };
-
-  const fetchShiftList = useCallback(async () => {
-    if (isLoading || !hasLoadMore) return;
-    setIsLoading(true);
-    try {
-      const response: IShiftList = await getShiftList(
-        page,
-        10,
-        workType,
-        approvalType,
-      );
-      const { currentPage, data, totalPage } = response;
-      setShiftList(prevList => [
-        ...prevList,
-        ...removeDuplicates(prevList, data, 'shiftSn'),
-      ]);
-      if (currentPage + 1 >= totalPage) setHasLoadMore(false);
-      return response;
-    } catch (error) {
-      throw new Error(`error: ${error}`);
-    } finally {
-      setIsLoading(false);
-    }
-  }, [isLoading, hasLoadMore, page, workType, approvalType]);
-
-  useEffect(() => {
-    fetchShiftList();
-  }, [fetchShiftList]);
-
-  const handleScroll = () => {
-    if (hasLoadMore) {
-      setPage(prev => prev + 1);
-    }
+    resetList();
   };
 
   return (
@@ -82,7 +42,7 @@ const ShiftPage = () => {
       <ShiftList
         isLoading={isLoading}
         listItem={shiftList}
-        onLoadMore={handleScroll}
+        onLoadMore={loadMore}
       />
     </S.Container>
   );
